Allow individual lines of code to be flagged for exclusion
This allows individual lines within a source file to be flagged for
exclusion.  Rather excluding the line from testing, a filter is applied
when aggregating results and any issues associated with flagged lines are
not included in output. The default flag is '# nosec', but this can be easily
customized or extended in bandit/constants.py.