Wildcard injection tests crash on non-string args
The wildcard injection tests only expect string call arguments.
This causes an AttributeError when we attempt to access the 's'
attribute of the call argument.

This patch first checks if the 's' attribute is present.  If it
is not present for an argument, we skip over that argument since
we won't be able to check if it's a wildcard anyway.  I also added
a function call to the wildcard example that triggers the crash
with the non-patched code.