Adding file discovery and directory exclusion
Bandit has relied on piping the find command to build the test
bed.  This change introduces file discovery capabilities to Bandit,
along with the capability to exclude directories and specify which
file extensions should be scanned.

Excluded directories and file extensions are specified in the
config file.  This change also introduces a '-r' flag which
indicates that Bandit should discover files recursively in a
directory.

Change-Id: I5721f0d067431e8eba2d2c2bc8037560d7b2df71