Add list of Python values considered False
The list of False values can be used by tests to check if
values are False or True more reliably. If a value is in
FALSE_VALUES it is false and otherwise is true. This will
only work in cases were the values being checked are
constants. This also does not deal with the case where a
user-defined class instance evaluates to false. We couldn't
figure that out anyway.

See https://docs.python.org/2/library/stdtypes.html#truth-value-testing

Change-Id: Ie301287c2516d311487dece0c098a36837f872dd