Fixed nosec flagging
nosec flags are now detected at the end of any line (even within multi-line
statements), and that line only is skipped (no tests are run).
The nosec flag is detected in a comment at the end of a line, and the check
is both case insensitive, and ignores whitespace.
As this fix is in the testing logic, rather than the reporting logic, nosec
will be honoured for all reporting mediums.

Change-Id: I95940ea06827a567eb2acae7071914f6f535fe10
Closes-Bug: #1432012