Fixed issue processing files containing invalid python
An issue existed when processing python files with invalid content
because the SyntaxError was being caught too early. This lead to
the file in question being skipped (no scores recorded), but still
remaining in the list of scanned files.
When bandit tried to report the results of this test, this resulted
in the reporting module trying to access scores that did not exist,
causing an error to be thrown.
The exception is now being caught in the correct place, and the
file skipped correctly. This fix also highlighted a flaw in the
functional test code, which has been fixed as well. It is now
possible to run tests, without explicitly checking the resultant
scores, using the run_example(test) method. This is useful, for
example, when providing tests that check files are skipped.

Change-Id: I0afcba912fba77353a34c2c4fb707c1001f1e5b4