Update README.rst
The README doc was out of date in a couple of regards.  It still
showed the old way of calling bandit with 'find', rather than the
new way of just specifying a directory.  It also showed the old
style of decorators, rather than the new style which are just
marked with the AST node type that they check.

Change-Id: If32a7135c4ff4eb91506b5203f3f2af4f9b7935e