Adding verbose flag
This commit adds a -v, verbose flag to Bandit.  Previously
included and excluded files were listed in output every time,
which made the output long and annoying.

This adds a -v flag to list the included and excluded files in
text mode output (for example) but it is passed to the instance of
result store, so output formatters can use it how they want.

Change-Id: I4ed44c5d2635f7dd1d465b6cf0942c7633ec7862