Fix config option fallback if "include" missing
Currently, if an "include" stanza is missing from the config
file and recursive checking is requested, a nasty error is
experienced.

When the included config option is used, it is supposed to
sanely fall back to the default of "*.py", but rather than
being set as a list (as is expected) a coding error sets it
to be a string.

This change resolves this issue by returning a list of one
item ['*.py'], rather than a string.

Change-Id: Ieb631032d1cc2d6e1864ca3d26b2155219fbd13e
Closes-bug: #1434281
Signed-off-by: Dave Walker (Daviey) <email@daviey.com>