Actually rely on entry-points for formatters
This should have been included in
I102277dcd9481f2573028a436e910eda10011d91 but somehow the code was
re-added in a later patchset. The way we picked our formatter function
prior to this was a bit too clever. Checking if the report_func is
self._report_text would not work and cause the format to always be 'txt'
even when an output file was supplied. To reproduce this try the
following:

    >>> class A(object):
    ...     def foo(self, bar):
    ...         pass
    ...
    >>> a = A()
    >>> b = a.foo
    >>> b is a.foo
    False

By properly using our extension manager, we can more simply check the
name of the formatter used and thus side-step this mess more easily.

Closes-bug: 1468075
Change-Id: I8e48cf40b2ef200b53e259ecf9ae54973b1aa0fb