Fixes exit code for filtered results
This commit fixes the return code when using a severity filter.
Prior to this commit the exit code would be '1' - indicating
results even when there weren't results displayed.  This was
because the exit code is based on total results whereas the
display is based on a filtered subset of the results.

Rather than filtering the results earlier on, this commit leaves
the full set of results available for formatters that may want it.
It also adds filter functionality to the results counter that the
main module uses to determine what the exit code should be.

Change-Id: I0b29ccab31c0898ebe84121d37cb0478ac3e2da9
Closes-Bug: #1480014
Closes-Bug: #1479216