Skip '/tests/' by default
This patch adds '/tests/' to the exclude of bandit.yaml by default.

That way, some projects can possibly get rid of their copy of
bandit.yaml completely by using a combination of severity and
confidence filters specified on the command line. Getting rid of
the bandit.yaml takes the work out of maintaining it every time
bandit pushes a new release to PyPi.

Change-Id: I2198c8c7fe7c6e1228aa28b390e4a23ea1ffb89d