Removing argument printing
Explicitly printing out the arguments to calls that have triggered a
warning is superfluous as the entire code string is printed out in
the message anyway. It is also error prone and fragile as the AST
nodes could be nearly anything, making it hard to decode properly.

This patch removes the explicit functionality.

Change-Id: I8e703cefc7c702e6fb876c3928f84a3f9c4410d9
Fixes-bug: 1497688