This adds baseline filtering to bandit
You can now provide bandit with a previous output report, in JSON
format. If given we will use this report to filter results. Results
will only be reported when the number of issues detected in a file
is not the same as the number that was detected in the baseline run.
If there is a discrepancy then all issues from the file will be
reported as we cant reliably determin the spoecific new issues
found.

Change-Id: Ia24ee665892ae59d7904776b9ff2637364ead213