Adding the plugin name to the HTML report
This commit adds the plugin name to the HTML formatter.  We're
building a process around developers checking the HTML reports and
referencing documentation to find more information about issues,
but without the plugin name they don't know what to reference.

We should consider adding the plugin to other formatters in the
future, but for now this adds it to HTML.

Change-Id: I15e74320e1cb8d29c49a38fa3282ff0b6c57ea2b