Fixing Baseline when a filter is used
This current baseline doesn't work if a filter is provided due
to the details mentioned here:
  https://bugs.launchpad.net/bandit/+bug/1512360

This fixes it by implementing the severity and confidence
filtering before baseline filtering.

Additionally, since each formatter used an IF statement to check
severity after getting issues, we've just moved the filtering to
the get_issues call.  If for some reason a formatter wants
to get unfiltered, they can omit the call.

Change-Id: I39c1d826f44a8c0d829fb7017d03dbff25779cba
Closes-Bug: #1512360