Making score sum totals more sane
This commit revamps the score sum system that was currently used
by JSON and TXT formatters.  The new approach is that an
aggregate SEVERITY and CONFIDENCE score is shown.

Previously the sum of all scores was added in a confusing way and
reported separately for each file in JSON metrics.  Text didn't
work at all, if you actually wanted to display them (with -v) it
crashed.

Change-Id: Id0c1f97795c6e1030b7fffcb1481da9c390dc4c1
Closes-Bug: #1512335
Closes-Bug: #1512333