Remove unnecessary absolute_import logic from modules
With bandit/bandit.py now living in bandit/cli, we do not need to worry
about using absolute imports inside of bandit because it will no longer
attempt to import the local bandit.py instead of the proper installed
package. (In other words it won't import bandit.bandit when it meant to
import bandit.)

We do need some of the absolute_import future imports in the formatters
and we now have comments explaining why for each of those.

Change-Id: Iceb8cc75ff8996e270c2992a0ee4a535a7f3827c