Make pep8 *the* linting interface
According to the PTI (=Python Test Interface,
http://governance.openstack.org/reference/cti/python_cti.html), pep8
is the interface for codestyle checks. Move all tests from linters to
pep8.

This change will be followed by a change to project-config to use pep8
for testing in the gate.

Note that the deps line is not needed, pep8 will use the default
environment.

Change-Id: I791b6f6e073e316e4e11867493647c917350eb4f