Adding profile generation to config generator
This enables config generator to create and save a profile
by embedding the possible plugin settings to a template which
includes some comments explaining the layout of the profile.

It is expected a user will delete the parts of the profile they
don't need and leave the parts they want.

Change-Id: Ie00aedfcf463685e0f15b793af9e2d9490a13180