Updated legacy profile support
This patch updates the logic that handles profiles in preparation
for deprecation. If no profile name is given then we use
"exculded_tests" and "included_tests" to read in the test filters
("exclude" and "include" were already used). If a profile name is
given then we will read in the filter as before.

In addition, this makes -s/-t/-p no longer mutually exclusive. If
-t/-s are used on the CLI then they will update the profile filter
read from the config (legacy or otherwise). This is handy for
adding in extra tests during a one-off manual run.

Eventually -p will be deprecated in favour of using -c to pick
from a set of new style config files.

This also adds -s/-t to the config gen tool.

Change-Id: Ibf701f68106195f0153a37303fb1bdf5a8c2df9c