Fixing an encoding error while writing HTML
While writing an HTML report for the ever-troublesome unit
test in Keystone:
  keystone/keystone/tests/unit/test_backend_ldap.py we generate
a UnicodeEncode error.  The solution is to re-encode UTF-8 prior
to writing a file.  In the case of HTML and TXT we'll do it on
file write.  In the case of JSON and CSV we'll do it when creating
the dict prior to write.

Closes-Bug: 1552326
Change-Id: I4ed18b5ad2b4b701ed6117e9aac42fe28feeb40f