Add functional tests for B308, B321, and B402
Several checks lack any functional test as described in the bug.
This patch adds mark_safe and ftplib tests.

There was also a typo in the calls doc where mark_safe was listed
under httpsconnection.

Also, the mark_safe check wasn't working because the full import
path for the call was not specified. That was also corrected.

Change-Id: I6f35fb65cb8c25a474175de99fcac04ea2b7d81e
Closes-Bug: #1648257