Add bandit ID to prefix of more_info link
In a recent commit [1], the names of the plugin doc files changed
to include the bandit ID as a prefix. Unfortunately, the doc_utils
wasn't updated at the time, so it still pointed to the previous
docs, thus resulting in 404 errors when browsing to the link.

This patch modifies doc_utils to properly prefix the bandit ID to
reference the doc for a particular plugin.

[1] https://review.openstack.org/#/c/540170/

Change-Id: Ia4b4c87e880ba39a677a84fc53943bc7a37849ef
Closes-Bug: #1761254