Repair some broken see also links in the doc
A few of the plugins used :doc: docstring tag to refer to another
doc. This doc style doesn't render properly when viewing the docs.
For example, take a look here:
https://bandit.readthedocs.io/en/latest/plugins/b502_ssl_with_bad_version.html#b502-ssl-with-bad-version

To fix this, I simply moved the link to the :seealso: block and
reference the other function using the :func: tag.

Signed-off-by: Eric Brown <browne@vmware.com>