Report dill usage
Dill is a wrapper around pickle and thus is vulnerable to the same
threats. Dill documentation states that dill "is not intended to be secure
against erroneously or maliciously constructed data. It is left to the
user to decide whether the data they unpickle is from a trustworthy
source." (https://pypi.org/project/dill/)