Re-eanble functional tests as part of CI
Currently our CI is only running the unit tests of Bandit found
in tests/unit/*.  However, much of the testing of Bandit occurs
as part of the functional tests in tests/functional/* that uses
the examples to test our many of the plugins.

This patch changes the stestr to use a test path in a parent directory
of unit and functional so that both are exercises as part of the CI
run. This will effectively run against each version of Python too.

http://stestr.readthedocs.io/en/latest/MANUAL.html#configuration

This patch also repairs the currently broken functional tests.
Namely:

* ast.TryFinally and ast.TryExcept were removed in Python 3.3+
  and replaced with ast.Try - Issue #326
* ast.Call kwargs argument was removed in 3.5 and replaced with
  keywords - Issue #350
* ast.Call starargs argument was removed in 3.5 and replaced with
  Starred in args - Issue #351
* ast.NameConstnat is used instead of ast.Name

Fixes Issue #326
Fixes Issue #350
Fixes Issue #351
Fixes Issue #353

Signed-off-by: Eric Brown <browne@vmware.com>