Properly handle nosec strings in code
The string "# nosec" has a chance (albeit remote) that it can be
part of the code on the line not just within the comment. Bandit
currently checks the entire line for that string.

This patch tokenizes the line of input to figure out what piece
is a comment, then compares whether that comment contains "# nosec".

Fixes #383

Signed-off-by: Eric Brown <browne@vmware.com>