Do not consider issue text in `__eq__` method
Issue text is not supposed to be consistent between banndit versions. It
was changed few times in the past. That's why we don't need to match text too
between issues.

This issue is only valid if we compare old bandit baseline generated
with the old version with a new one generated using the last bandit.