Remove leftover openstack code
There are still remnants of openstack in the Bandit code base.
Namely there was still a coverage step that relied on a openstack
git repo. Bandit is self contained and no longer only part of the
openstack ecosystem. Therefore it's safe to remove this coverage
step (which wasn't run anyway).

Signed-off-by: Eric Brown <browne@vmware.com>