Better hashlib check for Python 3.9
In Python 3.9 and later, the hashlib function has a new keyword
argument usedforsecurity to describe the usage of the hash. In
that way, we can better identify the severity of the error.

Previously, hashlib.md5 and the like were part of the blacklist
check. For Python 3.9, it'll be part of the hashlib plugin so
it can do more advanced checking of usedforsecurity.

Signed-off-by: Eric Brown <browne@vmware.com>