Use a table format for list of plugins
Rather than showing the plugins as a simple list, this change
puts them into a table. It also clearly separates active vs.
removed plugins for clarity.

Also discovered as part of this change was that snmp_weak_cryptography
plugin doc file was incorrectly named with the duplicate ID of 508
instead of 509 as it should be.

Signed-off-by: Eric Brown <browne@vmware.com>