Avoid calls to ast in plugins
If one day we do wish to have the capability to swap the builtin
Python ast for another parser, it's best to abstract any direct
calls to the ast parser in plugins.

This change addresses calls from plugins and other functions to
isinstance with ast node types as arguments. Further work is needed
for other ast calls in subsequent patches.

Related to #894

Signed-off-by: Eric Brown <eric_wade_brown@yahoo.com>